Screen Size, Game Table Cards
sorted out the issue with drawing on the bottom of the screen. Also got
the scroll bar, game hand and card layout working nicely.
Tied it all up by getting the networking up and running for adding cards
to the main table area. More work needed but now something that can be
built upon.