Pinch Zoom
Multi-touch is now in the game and just about working. The problem is
perspective needs to be altered after the rescaling to make any camera
tracks match the new zoom value.

Added problem of the positioning of the table not working as I expect
with no ability to track the table without the camera going completely
haywire.

All of which need to be fixed of course.