Pinch Fix and Table Grid
Figured out what was causing the panning issue in the load method (the
"tracking = yes" var in the border markers) and re-centered the table so
it spawns in 0,0. All works perfectly now.

Also introduced an update method to perspective to fix the centre of
screen issue. Now it's pretty much there.

Introduced a visual to display the table grid and show spacing.  Very
fine mesh at the moment. May want to adjust it so just the
centre-of-card markers are showing.