Zoom Fix and Moved Table and Grid
Figured out reading the perspective documentation how to move the camera
with no timing delay. This allows me to set this timing value to zero
and automatically position the camera at the start of the game and when
the pinch zoom works.

Also learned you can't have images more than 2024 x 2024 in size (or
near there anyway) so wasn't loading properly on the tablet. I changed
the size of the texture and everything now works.