Texture Size Fix, Checkboard Functionality
The large size of the table texture exceeded the 2000x2000 max so I had
to redesign how the table worked. It now can hold multiple table
elements that can be played next to one another.

Now includes the basic code for me to tell which marker positions should
contain card data when a card is moved.

Also includes the beginnings of a better hand functionality. e.g. the
hand popping back up when your selected card reaches a low point on the
screen.