Card Replacement, From Hand and Table
Two slightly different methods at the moment for recognising cards
that're dragged to the bottom of the screen and therefore available to
place back in a player's hand.

Also contains the beginning of a reset that can be used to remove more
than one card on the table that hasn't been finalised yet.