Just One - Current Card Check Fix
Fixed the coding needed to return the current card to the players hand
if they drop down another card. This means only one card can ever be
played and finalised (which is the current setup the the rest of the
game).

Included a remove current card method that updates the lists, images and
camera. Also included a "return to hand" method for the current card.