Hand Popup Fix and Face-Off Screen Initial Work
Fixed the issue of the hand not being able to pop up at certain zoom
distances and on differing screen sizes.

Sectioned off the load_setup stuff for the game.

Started on the ground work needed to start using a "face-off" popup in
the game. It's now to the point where I lock-off card functions and tie
their inclusion depending on an internal gamestate. This internal state
can be used when player's face-off, draw from decks or are playing the
main game.