Grid Amendments and Quad Implementation
Brought in the quad code I designed in the test repo. Modified it
slightly to create a list of classes that score the section number and
the card filename in the same location.

I can eventually alter this to have embedded sub lists of quads within
quads.

Also wrote a method to loop through the surrounding 4 directions of any
card grid position.

All I have to do now is figure out how to capture an accurate card
position when the card transitions from a card held in the hand and one
that's been transferred to the unsent "table list" area. While this
isn't resolved it's giving a false location for the card when it's first
placed on t he table.