Grid Check Implementation
I fixed the issue I was having trying to pull back the correct
positional information for clashing cards. The system still needs to
have card special rules and a few other things included but it does
basically what I need it to do: if a clash is deemed "not allowed" the
game hides the finalisation button.