Player Order and Network Recognition
The game was still missing a crucial element; the recognition of players
within the game. The networking code allowed two players connected to
the same room then see card movements after the connection but wasn't
done in such a way to say "this is player 1, this is player 2".

This update now had this coding