Bug Fixes and End Turn Function
Fixed an annoying bug that was stopping me replace a non-finalised card
on the table after the first turn is over.

Also isolated the end turn code so i'm not copy pasting code in the
networked and non-networked area of the game.

Included a "attacker/defender" value that'll be used to half the number
of action types i need to keep track of. All down to being able to
toggle who that actions directed at in a seperate value.