Card Abilities
After a long frigging night of trying to get it to work, I got the
"should be very simple to implement" card functions to actually
function. -sigh-

I've now covered health, armour, and both limb alterations. This
actually affects the stats of the player as well.

Play and draw card methods are the next thing to target, drawing cards
especially.