Draw Card Setup
Now that i've added the "play card" delay, the only ability left in
weapons to create was "draw card".

Decided I needed another screen to pop up which would allow the player
to select a deck to draw from, pull the card and place it in their own
hand.

The method is all set up so I now just have to write a method of drawing
actual card data, (maintaining that data until the draw card method is
cleared) then adding it to the players hand.