Networked Card Drawing and Character Classes
Put the method in that'll take it in turns to remove the card drawn by
each player, in each players deck. It does this by using the player list
to figure out which player gets to send data first.

Because of this I also thought it time to put in character class
information as I'd like to test the above methods are working correctly.