Draw Card Game Integration
Added an action list and state controls that allow me to cue up a number
of actions and have them play out in the main game.

There's also currently an exception used to trigger "play card" actions
but I think this can be rewritten in future.

May possibly have to consider reordering the actions list after it's
been added to gameinfo but think that can be avoided with decent action
maintenance.