Discard Card Game Integration
Added the discard card screen.

Also fixed the ability to advance the current actions being checked.
Because actions may apply to either the attacker or defender I had to
figure out a way of checking both types using the same method. Also
includes a networking solution to make sure both players are running
though the same actions (who it applies to is checked when the action
list is created)