End Round and Blocks Function
Introduced a means to check card action details in surrounding cards,
instead of just strat values. This is now where the block functions are
being checked. Included a card_type_int to card details as well so the
comparison is easier.

End round functions have been included and are largely there. Used to
reset card and board values as well as draw a new set of cards.