Wound Limb Table
The game now includes the ability to wound a limb. The major element
about this is that it has to use a networked method to ensure both
players have the same damage inflicted on limbs.

The methods really need a "all done now", you can swap players now
action. This'll be good as it'll take the "play" method out of the end
turn method and back into the main loop, where all the other methods
have been kept.

Next update should include the "waiting for other player" screen to pop
up. Just to know why you currently can't play an action (especially
useful if you're waiting for the opponent to play).