Shrapnel and Steal Actions
After playing some more of the game over the weekend, I decided to spend
a bit of time banging out the card lists so i can get to the real work
of making actions work.

Having done this I introduced Shrapnel and Steal actions; both of which
need the pass over information on the network.