Run Actions Bug and Placing Card Fixes
Through a bit of studying the old code I realised I ended up deleting
the "add to quad" code from the end turn function, which stopped the
table check method.
After that it was then making sure the "pass turn" function actually
reset the action position information and cleared the action list after
it passes the turn to the other player.