Face-Off Table Integration
Got the Face-off table working in terms of holding a card for the first
and second player within 2 zones. it also includes a method to put the
dropped card back into the hand under the normal circumstances used in
other actions.

I need to now change the finalisation button functions so the same
button can be used to submit the card for the faceoff.

I also need a couple of methods that'll allow me to display both cards
over the network.