Finalisation Button Modification -Part 1
Modded the finalisation button so I can now set a state that should (in
the future) work as a means of using the finalisation button for
differing functions.

Now i've got things setup to pass faceoff card info. I need to also make
sure that the finalisation button doesn't disappear for the "non current
player".