Finalisation Button Modification - Part 2
Put the show faceoff table code at the end of the network initiation
code, so it generates after the players have cards and after the game
elements are are set in place.

I can now also store faceoff cards in the player list and send the
filename to each player using the warplistener.

All I need to do now is generate the card on both ends, check each
player has the faceoff data, say who wins then go back to the rest of
the game. Part 3 is looming.