Finalisation Button Modification - Part 4
last changes to the finalisation button but still one or two more things
to do on the faceoff screen. I've got the faceoff screen to decide who
goes first as well as having it set up the faceoff action to be played.

Also prompted me to sort out the end_turn method again but it wasn't
actually working. Now I've managed to sucessfully end a turn on a
faceoff and restart with a faceoff screen that determines who wins
(something that proved a little but tricky).

One of the major things to remember is any action I intend to run on
both players through the main loop do not need to be sent over the
network.