Prev Card Instant Actions
Added method to check abilities module that allows me to declare that
the current card is going to be using values from the previous card to
complete an action. Both damage and armour can now be gained from the
previous cards pwoer value.

With a modification I could run the same method but just for the
previous card. This'll allow any "applies to next card" mod methods can
be run in the same place as well.