Counter Actions - Part 1
Introduced a pause in the end action phase that activates a timer for
the opponent (5 seconds i think) then send a "complete placement"
network message that, in turn, finishes the card placement for the
attacker.

its at the end of this timer and when the placement message is received
that the end_turn action stacks are created. Meaning they can now be
interrupted during this period.

From here I can introduce a counter ability for the opponent during
those 5 seconds, a card replacement method (should be easy) and a skip
to prevent the stacking of those now defunct actions.