Counter Actions - Part 3
This was the final ends of the counter actions and the bit that actually
had some feedback into the game. The method now gets rid of the last
card and adds a counter card in its place, ready to be processed by both
players.

Had a few little bits on the way bit hopefully they should be few and
far between and possibly just network errors.

Also got rid of the portrait screen element and started displaying the
enemies stats at the top of the screen and your own characters stats in
the scroll bar. These aren't set but at least demonstrate that both
players information needs to be visible on the screen at the same time.