Counter Action Text and Steal Fix
Managed to get the text to properly check at the end of a counter
action. Now the finalisation button describes the correct actions.

Also fixed the steal action as it wasn't showing the correct number of
cards stolen, nor was it moving to the next action. May have to check
there's no more of these.