Finalisation Fix and Limb Affect
The finalisation button now doesn't appear when your turn starts so you
have to place a card down in order to get it on the screen.

Limb's are now needed to place cards on the table or in faceoffs. Also
included popups in the card placement check so you know why a card can't
be placed.