Copycat - Part 1
I used map scrolling and card functions in order to select any finalised
card on the board. At the moment, it produces a popup that prints the
card filename.

Can't quite figure out where the actual copycat method should apply.

Turned off camera zooming as it's faulty. Needs a real look at.

Also allowed the camera to zoom back to the current card when it's
finalised.

Should maybe create a finalisation function that allows players to
refocus the camera when they've been moving around the board.