Copycat - Part 2
Written a method to hide the current card then add the selected card in
the same way any other card is created in the game.

Need to drop the whole "check strat points" as its a bit too odd to use.
Instead copycat will be a copy of rules only. Need to have a means of
replacing the strat points with a * symbol.

The counter point doesn't quite work when the card is added. Not sure
why this is. Also can't seem to finish the actions on the new card so
I'll have to check the source of that and fix it before i can call this
action complete.