Copycat - Part 3
I couldn't figure out why copycat was going wrong so i turned it all
off, gradually adding in the code again, bit by bit until i tracked down
the issues.

One problem was caused by not adding the card to the current players
table before sending the "add card update" that comes when the
finalisation button is set to 1. This meant that the counter switch was
being set to 1 for both attacker and defender when usually it should
only trigger for the defender. This meant two sets of actions where
being completed on one side of the game.

The second problems came to placing cards after the copied card was
placed. This i've tracked down to the current quad method that has no
hander for trying to add a quad where an existing quad resides. This
meant that no quad data was being returned when the quad_check was in
action so it couldn't check the last card.

Need to add a remove from quad method and possibly a multiple quad
method.