Faceoff Win Actions
Added in the means of deleting the end of an action stack if a faceoff
fails for the current player. This can be used for all card types.

Added in a ability to delay health mods so they can be considered in
faceoffs. Also added the same means for dealing faceoff power damage.

All the faceoff actions tied to cards have now been programmed in.