Pre-Game Actions and Flurry Damage
Included a select statement in the main loop of the code. This'll allow
me to stack methods together that can run as a constant loop within the
main loop. Should allow for more flexibility, especially now i'm
attempting to work in "pre-game" actions that need to run before the
main round takes place.

Something in the code above seems to stop the proper placement of cards
for the right place after a second round faceoff. Thought it might also
be the end round function but I tested it with the above turned off and
everything seemed to function correctly.

As a small aside, the game now finally includes a bloody flurry actions
with accompanying functions added to the finalisation button.