Or Actions #1
Created an "or action" table that allows me to generate a variable
number of buttons that can have a different type of action mapped to
them.

When the action button is selected, it inserts that action into the
current action list and advances the turn.

Just need to attach a warplistener that'll allow me to automatically
adjust heath.