DeathCheck #1
Added a method to check the amount of health of each player and the
number of cards in each deck. The method now needs to stop the game;
prevent other parts of the game from continuing and potentially have the
deck and health methods seperated, as both being true will never happen
as the game currently stands.