DeathCheck #4
Applied the end_game function to GameInfo and have now put in clauses to
stop the game from running when it's triggered. It should be quite easy
to impliment the reset now.

One remaining thing though is that the enemies life doesn't seem to
update when the clause triggers. Odd as it's still triggering the death
though. Should be easy to track down.