DeathCheck #5
The hunt for the reset bug continues. I've not remembered to pass
usernames forward once the reset occurs, but that hasn't fixed the
issue.
The bug seems to be based on the fact the second player isn't receiving
the message to draw a set of cards for himself in the warplistener.