RoundCheck #1
Put a good chunk of the code needed to run an end_round check in the
deathcheck class. It loops through the current players hand and runs a
quad_check on all the visible parts. If no card can be placed, a switch
should trigger to end the round.

All I ahve to do is create a method for checking around the positions
surrounding the last placed card, as these are the only positions the
hand cards can be placed in.

If there's no positions to place a card, this should also trigger the
end round method