Heal Limb Round Action
Actually turned out to be a bit more tricky than i first imaged as a new
table needed to be created for it to avoid action stacking issues, and
it uncovered a bug which meant that healing limbs couldn't be applied to
the attacking player; limb affects only applied to the defender.

Also contains numerous bug tweaks.