Network Rework #2
Centralised all appwork network messages to a single method. This'll
make it a lot easier to keep a message number and therefore be easier to
see if messages are being lost in the network.

If messages ARE BEING LOST then the same class can contain a queued
message method that'll have it's own state select method to send and
check for "message complete" strings.

Fixed the game reset as well. The problem was down to the character info
not being loaded onto the reset player after the game resets.