Network Rework #3
Added in a method that sends a "message received" response when a user
retrieves a message that isn't there own and isn't another "message
received" message.

I've also made sure that these messages don't contribute to the overall
message number.

This should allow me to track down if the issue i'm having is network
related.