End Round "Net Card" Bug
Noticed that when net was used in certain situations the game didn't
reset properly.

It took fucking ages to track down because it only seemed to crop up on
very rare occasions. After looking deeper into it, I managed to isolate
the problem using screenshots of the game working and the game when it's
broken.

The issue seemed to related to hand cards being present when they
technically shouldn't. This I think was because the set cards functions
was set to only playerlist[1] instead of using the current_player value
in gameinfo. This seems to be confirmed by adding a print function to
the code and checking it before and after the fix is implimemented. When
the fix is turned off, the hand check keeps looping instead of resolving
quickly as it normally would. Also checked that the first post and the
current_player pos brought up different usernames during the check,
which it did.