Card Animation States #1
There needs to be a consideration that all cards are going to have a
animation or message being played for each action on it. Therefore there
needs to be a pause between each action so this animation can be played.

That meant moving health, life and limb direct damage into the action
loop and introducing more functions the warplistener. I need to go
through these eventually and see if there's any duplicates.

I've also added a currently non-functioning animation loop that allows
me to at least pause between each action message. Meaning they're no
longer running all in a blink of an eye.