Card Animation States #3
I think I figured out the answer to the bug. The actions advancements
were being initalised in other players at the end of the current players
actions loop. This meant that if the current player were to count down
and send two messages before the first is properly complete by the
second player, the actions aren't advanced properly.