End Game Messages #1
I wanted to make sure that the proper end game messages were showing as
it's sometimes been unclear where the game is working properly or not
without it.

I've now altered the reset method so it doesn't clear messages; the end
turn method doesn't clear the end game message with a health delay
update.

I now need to check that the same happens when the round ends and when
the game ends through a player no longer having any life left.