Cripple / Heal Limb Text #1
Added text to say what the buttons represent. Also have them gray out if
there's less limbs on the player than buttons.

Need to remove the "button pressed" functionality, or at least prevent
it from running, if the button is grayed out.

Also need "Heal Arm/Leg" text on buttons when the function is used to
heal a limb.