Bug Fix - End Turn Major Glitch
I was having a problem with the card "sabotage" not resolving properly.
The main issue was that for the most part, actions would complete on one
side of the network but not on the other. The baffling thing though was
that it worked occationally.

Tracked the issue down to the end round function. When the lead player
passes over damage, they're being able to resolve their side of the end
turn reset method seconds before the other side. What this means is the
first player is checking the second players non-reset cards and assuming
they've actually gotten their cards for the next round.

Got past the issue by reusing the "reveal screen" method I used to check
that both players are ready to use the select player screen.