End Round (DeathCheck) Bug Fix
The networking delay issues came up again during the deathcheck
end_round method. The reason for this was, unlike "end turn" actions,
this method is initiated by one player, which causes the delay.

Ultimately the next faceoff couldn't start because the "nope, haven't
drawn cards yet, or am in a position to" check wasn't set up during the
method so wasn't preventing the reset from occuring as it now does
elsewhere.